==> 2_Netflix_RS_format_train_test_sets.ipynb <==
import warnings
warnings.filterwarnings("ignore")

from datetime import datetime

import numpy as np
import torch
import pandas as pd

import random
import collections

from progressbar import ProgressBar   

# DATA:

path = './data/'

%time df = pd.read_csv(path +'data.csv')
df.columns = 'movie user rating date'.split()

df.head()

df = df.iloc[:, :-1]
df.head()

listO_movies = df.movie.unique()
n_movies = len(listO_movies)
n_movies

listO_users = df.user.unique()
n_users = len(listO_users)
n_users

n_ratings = len(df)
n_ratings

# TRAIN & TEST SET FORMATTING:

### 1. Sample 5000 movies:

movies = list(range(1, n_movies+1))
n_samples = 5000

sampled_movies = np.random.choice(movies, n_samples, replace=False)
np.save(path + '5000_sampled_movies', sampled_movies)

sampled_movies = np.load(path + '5000_sampled_movies.npy')
sampled_movies

def retrn_dataset_sampled_by_movies(df, sampled_movies):
    
    from progressbar import ProgressBar
    pbar = ProgressBar()
    
    def retrn_tensorO_movie(movie):
        matrixO_movies = df[df.movie == movie].values
        tensorO_movie = torch.LongTensor(matrixO_movies)
        return tensorO_movie
    
    
    tensorO_data = torch.LongTensor()
    anamoly = []
    
    sampled_movies = pbar(sampled_movies)
    for movie in sampled_movies:
        
        try:
            tensorO_movie = retrn_tensorO_movie(movie)
        except:
            anamoly.append(user)
            
        tensorO_data = torch.cat((tensorO_data, tensorO_movie)) 
    
    return tensorO_data

sampled_movies = sampled_movies

tensorO_data_sampled_movies = retrn_dataset_sampled_by_movies(df, sampled_movies)

tensorO_data_sampled_movies

len(tensorO_data_sampled_movies[:, 0].unique())

len(tensorO_data_sampled_movies[:, 1].unique())

tensorO_data_sampled_movies.shape

np.save(path + 'tensorO_data_sampled_movies', tensorO_data_sampled_movies)

***

### 2. Sample 50,000 users from dataset created above:

users = tensorO_data_sampled_movies[:, 1].unique()
n_samples = 50000
sampled_users = np.random.choice(users, n_samples, replace=False)
np.save(path + '50000_sampled_users', sampled_users)

sampled_users = np.load(path + '50000_sampled_users.npy')
sampled_users

sampled_users.shape

df_sampled_movies = pd.DataFrame(tensorO_data_sampled_movies.numpy(), columns = 'movie user rating'.split())
print(df_sampled_movies.shape)

df_sampled_movies.head()

def retrn_dataset_sampled_by_users(df, sampled_users):
    
    from progressbar import ProgressBar
    pbar = ProgressBar()
    
    def retrn_tensorO_user(user):
    
        matrixO_user = df[df.user == user].values
        tensorO_user = torch.LongTensor(matrixO_user)
        return tensorO_user
    
    
    tensorO_data = torch.LongTensor()
    anamoly = []
    
    sampled_users = pbar(sampled_users)
    for user in sampled_users:
        
        try:
            tensorO_user = retrn_tensorO_user(user)
        except:
            anamoly.append(user)
            
        tensorO_data = torch.cat((tensorO_data, tensorO_user)) 
    
    return tensorO_data

df = df_sampled_movies
sampled_users = sampled_users

sampled_dataset = retrn_dataset_sampled_by_users(df, sampled_users)

sampled_dataset

sampled_dataset.shape

len(sampled_dataset[:, 0].unique()), len(sampled_dataset[:, 1].unique())

np.save(path + 'sampled_dataset', sampled_dataset)

***

### 3. TRAIN-TEST SPLIT:

df_sampled_dataset = pd.DataFrame(sampled_dataset.numpy(), columns = 'movie user rating'.split())
print(df_sampled_dataset.shape)

df_sampled_dataset.head()

def shuffle(df):
    index = list(df.index)
    random.shuffle(index)
    df = df.ix[index]
    df.reset_index()
    return df

df_sampled_dataset = shuffle(df_sampled_dataset)
df_sampled_dataset.head()

limit = int(0.8 * len(sampled_dataset))

df_trainset = df_sampled_dataset.iloc[:limit, :]
df_testset  = df_sampled_dataset.iloc[limit:, :]

df_trainset.shape, df_testset.shape

df_trainset.head()

df_testset.head()

### 4. Filter out those users from testset which are not preset in trainset:
- AVOID COLD START

unique_train_users, unique_test_users = df_trainset.iloc[:, 1].unique(), df_testset.iloc[:, 1].unique()

unique_train_users.shape, unique_test_users.shape

new_users = list(set(unique_test_users) - set(unique_train_users))
len(new_users)

def fn_remove_new_users(df_testset, new_users):
    
    pbar = ProgressBar()
    
    for new_user in pbar(new_users):
        df_testset = df_testset[df_testset.user != new_user]
        
    return df_testset

df_testset = fn_remove_new_users(df_testset, new_users)
df_testset.head()

df_testset.shape

### CHECKING IF TESTSET CONTAINS ANY 'NEW' USERS :

unique_train_users  = df_trainset.iloc[:, 1].unique() 
unique_test_users = df_testset.iloc[:, 1].unique()

new_users = list(set(unique_test_users) - set(unique_train_users))
len(new_users)

***

### 5. Embed users & movies as continuous numbers:
- Pytorch Embedding layer expects numerical IDs to lie withing the range of the Embedding dimension size specified
- ie: If n_users = 5000, then all unique user ids should have values between (0, 5000)

def fn_embed_id_cols(df_train_set, df_test_set):
    
    pbar = ProgressBar()
    
    df_trainset = df_train_set.copy()
    df_testset  = df_test_set.copy()
    
    columns = pbar(df_trainset.columns[:-1])
    for col in columns:
        unique = df_trainset[col].unique()
        
        look_up_table = {j : i for i, j in enumerate(unique)}
        
        trainset_col = [look_up_table[j] for j in df_trainset[col].values]
        testset_col  = [look_up_table[j] for j in df_testset[col].values]
        
        df_trainset[col] = trainset_col
        df_testset[col]  = testset_col
        
    return df_trainset, df_testset


df_trainset_encoded, df_testset_encoded = fn_embed_id_cols(df_trainset, df_testset)

df_trainset_encoded.head()

df_testset_encoded.head()

df_trainset_encoded.shape, df_testset_encoded.shape

### CHECKING IF TESTSET CONTAINS ANY 'NEW' USERS :

set(df_testset_encoded.user.unique()) - set(df_trainset_encoded.user.unique())

### SAVE:

df_trainset_encoded.to_csv(path + 'df_final_trainset.csv')
df_testset_encoded.to_csv(path + 'df_final_testset.csv')

***
