==> 4_Netflix_MatrixFactorization_regressn_Model_2.ipynb <==
from google.colab import drive
drive.mount('/content/gdrive')

# http://pytorch.org/
from os.path import exists
from wheel.pep425tags import get_abbr_impl, get_impl_ver, get_abi_tag
platform = '{}{}-{}'.format(get_abbr_impl(), get_impl_ver(), get_abi_tag())
cuda_output = !ldconfig -p|grep cudart.so|sed -e 's/.*\.\([0-9]*\)\.\([0-9]*\)$/cu\1\2/'
accelerator = cuda_output[0] if exists('/dev/nvidia0') else 'cpu'

!pip install -q http://download.pytorch.org/whl/{accelerator}/torch-0.4.1-{platform}-linux_x86_64.whl torchvision
import torch

import warnings
warnings.filterwarnings("ignore")

***

# START:

import numpy as np
import pandas as pd

from collections import Counter

import torch
import torch.nn as nn

from torch.autograd import Variable
from torch.utils.data import Dataset, DataLoader

### DEVICE:

device = 'cuda' if torch.cuda.is_available() else 'cpu'
device

### PATH:

path='gdrive/My Drive/netflix_colab/data/'

***

## FUNCTONS:

### 1. FN FOR TRAINING & VALIDATION:

def fn_train_eval(model, train_dataset, test_dataset, batch_size, optimizer, criterion, n_epochs):
    
  
    # INNER FUNCTIONS-----------------------------------------------
    
    def fn_train(model, iterator, optimizer, criterion):

        epoch_loss = 0
        model.train()
        for batch in iterator:
            
            movies  = Variable(batch[0][:, 0]).to(device)
            users   = Variable(batch[0][:, 1]).to(device)
            ratings = Variable(batch[1]).to(device)
            
            optimizer.zero_grad()                   # INITIALIZE
            
            predictions = model(users, movies)      # PREDICT
            loss = criterion(predictions, ratings)  # COMPUTE LOSS

            loss.backward()                         # BACK PROP
            optimizer.step()                        # GRADIENT DESCENT

            epoch_loss += loss.item()

        epoch_loss = epoch_loss/len(iterator)

        return epoch_loss 


    def fn_evaluate(model, iterator, criterion):

        epoch_loss = 0
        model.eval()
        with torch.no_grad():

            for batch in iterator:
            
                movies  = Variable(batch[0][:, 0]).to(device)
                users   = Variable(batch[0][:, 1]).to(device)
                ratings = Variable(batch[1]).to(device)


                predictions = model(users, movies)
                loss = criterion(predictions, ratings)

                epoch_loss += loss.item()

        epoch_loss = epoch_loss/len(iterator)

        return epoch_loss 
    
    
    # RUN--------------------------------------------------------------------

    train_iterator = DataLoader(dataset=train_dataset, batch_size=batch_size, shuffle=True, num_workers=2)
    test_iterator  = DataLoader(dataset=test_dataset,  batch_size=batch_size, shuffle=True, num_workers=2)

    listO_train_losses, listO_test_losses = [], []
    
    c1, c2 = 'print', 'print'
    for epoch in range(n_epochs):

        train_loss = fn_train(model, train_iterator, optimizer, criterion)
        if c1 == 'print':
            print('TRAINING >>')
            c1 = 'no_print'

        test_loss = fn_evaluate(model, test_iterator, criterion)
        if c2 == 'print':
            print('EVALUATING >>')
            c2 = 'no_print'

        listO_train_losses.append(train_loss)
        listO_test_losses.append(test_loss)
        
        print_params = [epoch+1, train_loss, test_loss]
        print('Epoch: {:03}  |  Train Loss: {:.3f}  | test Loss: {:.3f}  |'.format(*print_params))
        
    print()
    print('RETURNED: listO_train_losses, listO_test_losses, model')
    
    return listO_train_losses, listO_test_losses, model 

### 2. PLOTTING FN:

def retrn_model_plots(listO_train_losses, listO_test_losses):
    
    import pylab
    pylab.rcParams["figure.figsize"] = (10, 5)

    xs = list(range(1, len(listO_train_losses)+1))
    ys = listO_train_losses
    ys_v = listO_test_losses


    pylab.plot(xs, ys, '-', label = 'train')
    pylab.plot(xs, ys_v, '-', label = 'test')

    pylab.legend()
    pylab.xlabel('EPOCHS')
    pylab.ylabel('LOSS')
    
    pylab.show()

***

### CHECK DATA:

df_train = pd.read_csv(path + 'df_final_trainset.csv', index_col = 0)
df_train.head()

df_train.movie.max(), df_train.user.max()

***

# CREATE - TRAIN & TEST PYTORCH DATASET CLASSES:

class RecommenderData(Dataset):

    def __init__(self, path, csv_file_name):
        
        df = pd.read_csv(path + csv_file_name, index_col = 0)
                        
        self.len = df.shape[0]
        
        self.x_data = torch.from_numpy(df.iloc[:, :-1].values).type(torch.LongTensor)
        self.y_data = torch.from_numpy(df.iloc[:, -1].values).type(torch.FloatTensor)

    def __getitem__(self, index):
        return self.x_data[index], self.y_data[index]

    def __len__(self):
        return self.len

path = path

train_dataset = RecommenderData(path, 'df_final_trainset.csv')
test_dataset  = RecommenderData(path, 'df_final_testset.csv')

train_dataset.x_data

train_dataset.y_data

train_dataset.x_data[:, 0].max(), train_dataset.x_data[:, 1].max()

train_dataset.len, test_dataset.len

***

# MODEL:

class MatrixFactorization_2(torch.nn.Module):
    
    def __init__(self, n_users, n_items, embedding_dim, hidden_dim, last_dim):
        super().__init__()
        
	    # create user & item embeddings of same size:
        self.user_embeddings = nn.Embedding(n_users, embedding_dim, sparse=False)
        self.item_embeddings = nn.Embedding(n_items, embedding_dim, sparse=False)
        
        self.weights1 = nn.Linear(embedding_dim, hidden_dim)
        self.weights2 = nn.Linear(hidden_dim, last_dim)
        
        self.batch_norm1 = nn.BatchNorm1d(embedding_dim)
        self.batch_norm2 = nn.BatchNorm1d(hidden_dim)
        
        self.dropout1 = nn.Dropout2d(p=0.5)
        self.dropout2 = nn.Dropout2d(p=0.25)
        
        self.relu = nn.ReLU()

    def forward(self, user, item):
        
    	# CREATING USER LATENT VARIABLES:
        # layer 1
        user_vec = self.user_embeddings(user)
        user_vec = self.batch_norm1(user_vec) 
        user_vec = self.dropout1(user_vec) 
        user_vec = self.weights1(user_vec) 
        user_vec = self.relu(user_vec)
        # layers 2 
        user_vec = self.batch_norm2(user_vec) 
        user_vec = self.dropout2(user_vec) 
        user_vec = self.weights2(user_vec) 
        user_vec = self.relu(user_vec)
        
        
        # CREATING ITEM LATENT VARIABLES:
        # layer 1
        item_vec = self.item_embeddings(item)
        item_vec = self.batch_norm1(item_vec) 
        item_vec = self.dropout1(item_vec) 
        item_vec = self.weights1(item_vec) 
        item_vec = self.relu(item_vec)
        # layers 2 
        item_vec = self.batch_norm2(item_vec) 
        item_vec = self.dropout2(item_vec) 
        item_vec = self.weights2(item_vec) 
        item_vec = self.relu(item_vec)
        
        y_pred = (user_vec * item_vec).sum(1)
        
        return y_pred   

***

# MODEL TRAINING & EVALUATION:

### TRIAL 1: Embedding_dim, hidden_dim, last_dim = 100, 50, 5

 n_users, n_items, embedding_dim, hidden_dim, last_dim = 50000, 5000, 100, 50, 5

model = MatrixFactorization_2(n_users, n_items, embedding_dim, hidden_dim, last_dim)

learning_rate = 1e-3
optimizer = torch.optim.Adam(model.parameters(), lr=learning_rate)
criterion = nn.MSELoss()

model = model.to(device)
criterion = criterion.to(device)

train_dataset = train_dataset
test_dataset = test_dataset
batch_size = 1000
n_epochs = 50

%time listO_train_losses, listO_test_losses, model = fn_train_eval(model, train_dataset, test_dataset, batch_size, optimizer, criterion, n_epochs)

***

### TRIAL 2:  Embedding_dim, hidden_dim, last_dim = 1000, 500, 100

 n_users, n_items, embedding_dim, hidden_dim, last_dim = 50000, 5000, 1000, 500, 100

model = MatrixFactorization_2( n_users, n_items, embedding_dim, hidden_dim, last_dim)

learning_rate = 5e-3
optimizer = torch.optim.Adam(model.parameters(), lr=learning_rate)
criterion = nn.MSELoss()

model = model.to(device)
criterion = criterion.to(device)

train_dataset = train_dataset
test_dataset = test_dataset
batch_size = 1000
n_epochs = 50

%time listO_train_losses, listO_test_losses, model = fn_train_eval(model, train_dataset, test_dataset, batch_size, optimizer, criterion, n_epochs)

***

### TRIAL 3: Embedding_dim, hidden_dim, last_dim = 500, 100, 10

 n_users, n_items, embedding_dim, hidden_dim, last_dim = 50000, 5000, 500, 100, 10

model = MatrixFactorization_2( n_users, n_items, embedding_dim, hidden_dim, last_dim)

learning_rate = 5e-3
optimizer = torch.optim.Adam(model.parameters(), lr=learning_rate)
criterion = nn.MSELoss()

model = model.to(device)
criterion = criterion.to(device)

train_dataset = train_dataset
test_dataset = test_dataset
batch_size = 1000
n_epochs = 50

%time listO_train_losses, listO_test_losses, model = fn_train_eval(model, train_dataset, test_dataset, batch_size, optimizer, criterion, n_epochs)

***

### TRIAL 4:  Embedding_dim, hidden_dim, last_dim = 500, 250, 5

 n_users, n_items, embedding_dim, hidden_dim, last_dim = 50000, 5000, 500, 250, 5

model = MatrixFactorization_2( n_users, n_items, embedding_dim, hidden_dim, last_dim)

learning_rate = 5e-3
optimizer = torch.optim.Adam(model.parameters(), lr=learning_rate)
criterion = nn.MSELoss()

model = model.to(device)
criterion = criterion.to(device)

train_dataset = train_dataset
test_dataset = test_dataset
batch_size = 1000
n_epochs = 50

%time listO_train_losses, listO_test_losses, model = fn_train_eval(model, train_dataset, test_dataset, batch_size, optimizer, criterion, n_epochs)

***

### TRIAL 5:  Embedding_dim, hidden_dim, last_dim = 500, 250, 15

 n_users, n_items, embedding_dim, hidden_dim, last_dim = 50000, 5000, 500, 250, 15

model = MatrixFactorization_2( n_users, n_items, embedding_dim, hidden_dim, last_dim)

learning_rate = 5e-3
optimizer = torch.optim.Adam(model.parameters(), lr=learning_rate)
criterion = nn.MSELoss()

model = model.to(device)
criterion = criterion.to(device)

train_dataset = train_dataset
test_dataset = test_dataset
batch_size = 1000
n_epochs = 50

%time listO_train_losses, listO_test_losses, model = fn_train_eval(model, train_dataset, test_dataset, batch_size, optimizer, criterion, n_epochs)

## BEST PERFORMANCE : TRIAL 3 - Epoch: 029  |  Train Loss: 0.722  | test Loss: 0.797
